# Faster Utils.clean_path_info