@matthewd 2017-08-25
Assuming this method is hot enough to care about, a couple of extra quick-glance observations:

* In the common case of a single path_sep, the split should be measurably faster with a string instead of a regexp
* I suspect this often has nothing to actually clean... could it be worth an early no-op return unless the original path matches `/#{PATH_SEPS}\.*#{PATH_SEPS}/`?