@schneems 2017-08-25
If we could early return than we could skip two array allocations, and that would be much faster. I'll throw that on here in a bit. I would want to use `Regexp#match?` so i'll need to add a backwards compat shim if it's not somewhere in here already. My only worry would be security. 

We could simply check for any presence of a dot `/\./` and we might have higher false positives for things like assets `foo.jpg` but I would be less worried about obscure edge cases.

For splitting via string instead of regex, we're using both the path separator and an alternate path separator. You can't split on two strings.