@matthewd 2017-08-25
> so i'll need to add a backwards compat shim

I think we have precedent somewhere that the compatibility would be in the form of a conditional around this method, defining two different versions based on whether the thing we need (`#match?`, in this case) is available.

> We could simply check for any presence of a dot

The algorithm's not that complicated... the only conditions inside the loop body are for an empty string, exactly `'.'`, and exactly `'..'`.

> For splitting via string instead of regex, we're using both the path separator and an alternate path separator. You can't split on two strings.

Right, but those only differ on Windows. On my laptop, and most server systems:

```
[::File::SEPARATOR, ::File::ALT_SEPARATOR].compact
=> ["/"]
```

So it could be a regex when they differ, and a string when they don't.