@schneems 2019-11-16
Rebased, I didn't end up implementing any of the early return optimizations as discussed earlier though. I would probably split off into a separate PR at this point.

Tests are failing. Not sure the status of CI on master.